LayoutTests/media/video-dom-loopcount-expected.txt

Test that playcount DOM attribute causes playback to end after right number of repeats.

EXPECTED (video.playCount == '3') OK
EXPECTED (video.currentLoop == '0') OK
EXPECTED (video.getAttribute('playCount') == '3') OK
RUN(video.play())
EVENT(ended)
EXPECTED (video.currentLoop == '2') OK
END OF TEST
